Fix WinRT reinitialization in Support module tests
